Allow mpris to play after queue is loaded, and after playlist ends
Sets mpris::can_play(true) after a playlist is first loaded, which allows
playback to be started from mpris without first clicking play in the application
window. The playback state is also set to 'pause' rather than 'stop' after
a playlist ends; this is a workaround for the currently lacking ability of
mpris to transition Amberol from the stopped to the playing state.

Closes #342