Add mpris.connect_volume and partly fix seek_position_abs bug
Enable updating volume through mpris (using KDE connect for example).
set_volume has been added too, but is is not called because a bug
remains that prevents connect_volume and set_volume from being used
at the same time. When both are used, they call each other in an
infinite loop which causes the system to freeze.

A logic bug in seek_position_abs had been seeking to the end of track.
It looks like the intent was to ensure that the seek position called
did not exceed the length of the track, so that is the logic I
implemented.

Another bug still exists in seek_position_abs, causing
random seeking. Similarly to set_volume, this bug can be avoided by not
calling set_position.