audio: emit "Seeked" MPRIS signal when needed
The MPRIS specification indicates that the "Seeked" signal should be
emitted when the track position changed in a non-continuous way (i.e. a
`Seek` action has been processed), and only in such cases.

In order to achieve this behaviour, the `UpdatePosition` action is
extended with an additional boolean parameter: when that parameter is
`true`, then "Seeked" should be emitted. This case is triggered when
receiving a `seek_done` signal from `gst_player`, as it indicates a
`Seek` action was previously executed.

When the position update is triggered by the `position_updated` signal,
then this parameter is `false`, so we avoid flooding MPRIS clients with
unneeded occurrences of "Seeked".